pyproject.toml
[tool.poetry]
name = "API"
version = "0.1.0"
description = ""
authors = ["Mikayel Ohanyan <mikaohanyan881@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-multipart = "^0.0.6"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.25"}
Jinja2 = "^3.1.2"
aiosqlite = "^0.19.0"
pydantic-settings = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
